Use macro expandee, rather than expansion, in pres. compiler
The presentation compiler is primarily interested in trees that
represent the code that one sees in the IDE, not the expansion of
macros.

This commit continues to expand macros, but adds a hook in which
the presentation compiler discards the expansion, retaining instead
the expandee. The expandee is attributed with the type of the
expansion, which allows white box macros to work. In addition,
any domain specific errors and warnings issued by the macro will
still be reported, as a side-effect of the expansion.

The failing test from the last commit now correctly resolves
hyperlinks in macro arguments.

Related IDE ticket:

  https://www.assembla.com/spaces/scala-ide/tickets/1001449#

This facility is configured as follows:

    // expand macros as per normal
    -Ymacro-expand:normal

    // don't expand the macro, takes the place of -Ymacro-no-expand
    -Ymacro-expand:none

    // expand macros to compute type and emit warnings,
    // but retain expandee. Set automatically be the presentation
    // compiler
    -Ymacro-expand:discard

This leaves to door ajar for a new option:

    // Don't expand blackbox macros; expand whitebox
    // but retain expandee
    -Ymacro-expand:discard-whitebox-only

The existing test for SI-6812 has been duplicated. One copy exercises
the now-deprecated -Ymacro-no-expand, and the other uses the new
option.

.