SI-7788 Predef.conforms can circumvent implicit shadowing
Regular implicits play by the rule that they are only usable
implicitly if they could be used explicitly. This means that
they cannot be shadowed.

`Predef.conforms` is a bit special: it already has one special
rule for implicit search, in that it can't be used for implicit
views.

This commit adds a second special case: it cannot be shadowed
by another non-implicit member.

Conflicts:
	src/compiler/scala/tools/nsc/typechecker/Implicits.scala