Improve shadowed implicits detection, reporting.
Rather than using `typedIdent` to detect shadowing, do a direct
lookup in the scope with `Context#lookup`.

In the process, we simplify the manner in which the implicit tree
is created, but delegating the job to `TreeGen`, after extending
it to support creating `Ident`s with a name different to the
referenced symbol, as can happen with aliased imports.

Before this change, the reason for shadowed implicit was obscured
by the inability to typecheck the shadower.

    % scalac -Xlog-implicits neg/t7788.scala
    conf is not a valid implicit value for T[String] because:
    hasMatchingSymbol reported error: polymorphic expression cannot be instantiated to expected type;

En route, I had to refactor the remainder of typedImplicit.