Fix Ant uptodate checking in OSGI JAR creation
Changes to the compiler or library were not triggering recreation
of the OSGI JARs, which are used in the local maven deployment tasks.

Reason: this Ant-ipattern:

    <fileset base="/foo">
       <include name="/foo/bar.baz"/>
    </fileset>

Instead, we should use a <filelist/>, which doesn't require
a base directory. This can be composed into a <srcresources/>
element in the <uptodate/>.

Phew!