SI-8006 prevents infinite applyDynamicNamed desugarings
Since mkInvoke, the applyDynamic/selectDynamic/etc desugarer, is disconnected
from typedNamedApply, the applyDynamicNamed argument rewriter, the latter
doesn’t know whether it needs to apply the rewriting because the application
has just been desugared or it needs to hold on because it’s already performed
a desugaring on this tree.

This commit introduces the attachment that links these translation facilities,
preventing infinite applyDynamicNamed desugarings.