remove symbol-based factories for MemberDef AST nodes
As recent discussion shows, methods like ValDef(sym) or DefDef(sym, rhs)
are not just useless (i.e. not achieving the goal that users expect them
to achieve), but they are also misleading (i.e. actively confusing users
from utilizing the reflection API the right way):
http://stackoverflow.com/questions/20908036/how-do-i-splice-symbols-of-various-types-in-a-quasiquote.

Therefore this PR removes these factories from the public API in 2.11.0.
Even though we don’t have a formal right to do that, as the factories
weren’t deprecated before 2.10.1, I strongly propose to go for it, because:
1) these APIs are actively harmful, 2) reflection is an experimental API,
so we’re not bound as tightly as usual.