SI-8125, SI-8130 Corrected position of synthetic `apply`, `applyDynamicNamed`
* Tickets SI-8125 and SI-8130 turn out to be caused by the same underlying
issue, i.e., sythetic `apply` Tree where created with the wrong kind of
position. The initial fix (look at the first set of changes in
Typers.scala) caused a side-effect in the kind of positions assigned to
`applyDynamicNamed` synthetic Trees, which changed from transparent to
opaque positions. This was fixed by assigning transparent position for
synthetic `applyDynamicNamed` Trees (see the second set of changes in
Typers.scala).

* For test `t8125`, note that `def apply(a: Int): Int` is part of the
result returned by the completion test.

* For test `t8130`, note that hyperlinking on `d` now is correctly resolved
to the declaration of `d`, and not the `applyDynamic` method in class `D`.