SI-8124 Scaladoc package object members
Modifies the package pruning test to check whether the
package was first seen during namer.  Scaladoc gets a
custom namer that remembers package symbols.

A nicer mechanism may have been to turn on symbol
tracking and then query the tracker after namer, but
there are no hooks presently for that.