SI-8143 Fix bug with super-accessors / dependent types
Super-accessors are generated as `DefDef`'s with `EmptyTree` as a
placeholder for the RHS. This is filled in later in `Mixin` in
`completeSuperAccessor`.

A change in `Uncurry` (SI-6443 / 493197f), however, converted this
to a `{ EmptyTree }`, which evaded the pattern match in mixin.

This commit adds a special case to the dependent method treatment
in Uncurry to avoid generating redundant blocks.