Avoid cycles in Symbol toString under -Ydebug
This is the first of two commits to restore workingness
to the compiler under `-Ydebug`.

`ResetAttrs` is now called during case class unapply synthesis,
after the UnTyper was recently banished.

But, this class has some low-level tracing that is triggered
under `-Ydebug` (irrespective of any `-Ylog` settings.)

This tracing code calls `Symbol#toString`, which, in an attempt
to discriminate primary from secondary constructors, accesses
the info of its owner. This is sufficient to hit a dreaded
`CyclicReferenceError`.

The enclosed test compiles a case class under this option
to show that things now compile. It still spews out unwanted
output; this will be removed in the next commit.