ensures that runtime reflection symbols are thread-safe
A long time ago, when implementing the first version of reflection sync,
we discovered that, even though we had the SynchronizedSymbol trait that
covered critical methods of Symbol with locks, there were no symbols to
benefit from that, because everyone’s transitively created by root and
root itself wasn’t synchronized :)

It looks like being all euphoric over discovering and fixing such a
ridiculous bug, we’ve missed some of the symbols which remained unsynced,
e.g. NothingClass. This commit fixes this oversight and enforces the policy
in an assertion that now lives in Symbol’s base constructor.