Make abstract-report2 test less dependent on std library.
Replace the use of `Ordering` by custom, dummy trait Xyz defined in test's
source.

By not inheriting from Ordering in abstract-report2 we make the test
less dependent on both Scala and Java library. The reason we are less
dependent on Java is because Ordering extends Java's comparator.

This change is motivated by the fact that Java 8 introduced default
method `reversed` to Comparator interface and we get a failure due to
conflicting inheritance:

-abstract-report2.scala:9: error: trait Bippy inherits conflicting members:
-  method reversed in trait TraversableOnce of type => List[(T2, String)]  and
-  method reversed in trait Comparator of type ()java.util.Comparator[T3]
-(Note: this can be resolved by declaring an override in trait Bippy.)
-trait Bippy[T1, T2, T3] extends Collection[T1] with TraversableOnce[(T2, String)] with Ordering[T3]
-      ^