Prohibit views targeting AnyVal
Library changes in Scala 2.10 mean that we are left with the
unfortunate situation of admitting:

    scala> "": AnyVal
    res0: AnyVal =

 We already have explicit checks in place to prevent views
 targeting `AnyRef`. This commit balances this out by prohibiting
 `AnyVal`, as well.

The enclosed test shows that this case is now prevented. If multiple
implicits views are applicable, the ambiguity error is still raised;
these check comes right at the end. Maybe that ought to be changed,
but I don't think it matters too much.

I've also disabled this prohibition under -Xsource:2.10.